task-driver: tasks: update-wallet: Modularize update wallet impl
This comes ahead of additional logic to precompute cancellation proofs.
Breaking the helpers into a module will help segment components of the
update wallet task into more discoverable locations.
