+++ workers/task-driver/src/tasks/update_wallet/mod.rs
@@ -1,0 +1,5 @@
+//! Defines the update wallet task
+
+mod helpers;
+mod task;
+pub use task::{UpdateWalletTask, UpdateWalletTaskError, UpdateWalletTaskState};
